+++ Panther/src/sema/SemanticAnalyzer.h
@@ -136,6 +136,17 @@ namespace pcit::panther{
 			EVO_NODISCARD auto instr_move(const Instruction::Move& instr) -> Result;
 			template<bool IS_READ_ONLY>
 			EVO_NODISCARD auto instr_addr_of(const Instruction::AddrOf<IS_READ_ONLY>& instr) -> Result;
+
+			template<bool IS_CONSTEXPR>
+			EVO_NODISCARD auto instr_prefix_negate(const Instruction::PrefixNegate<IS_CONSTEXPR>& instr) -> Result;
+
+			template<bool IS_CONSTEXPR>
+			EVO_NODISCARD auto instr_prefix_not(const Instruction::PrefixNot<IS_CONSTEXPR>& instr) -> Result;
+			
+			template<bool IS_CONSTEXPR>
+			EVO_NODISCARD auto instr_prefix_bitwise_not(const Instruction::PrefixBitwiseNot<IS_CONSTEXPR>& instr)
+				-> Result;
+
 			EVO_NODISCARD auto instr_deref(const Instruction::Deref& instr) -> Result;
 
 			template<bool IS_CONSTEXPR>